requirements.txt
selenium~=4.38.0
mysql-connector-python~=9.5.0
undetected-chromedriver~=3.5.5